+++ packages/core/src/config/payment-method/example-payment-method-handler.ts
@@ -26,8 +26,8 @@ export const examplePaymentHandler = new PaymentMethodHandler({
     code: 'example-payment-provider',
     description: [{ languageCode: LanguageCode.en, value: 'Example Payment Provider' }],
     args: {
-        automaticCapture: { type: 'boolean' },
-        apiKey: { type: 'string' },
+        automaticCapture: { type: 'boolean', required: false },
+        apiKey: { type: 'string', required: false },
     },
     createPayment: async (ctx, order, amount, args, metadata): Promise<CreatePaymentResult> => {
         try {